+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this package are documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v1.9.0] - 2023-11-08
+
+### Added
+- Added JsonConvert error handling support to `GameFile`
+  - Can be overriden by setting `GameFile.jsonSerializeSettings` and `GameFile.jsonDeserializeSettings`
+
 ## [v1.8.0] - 2023-10-10
 
 ### Added